Address IVF HNSW SQ review feedback
